Move the apiserver config structs to config.go
